api/requirements.txt
-e file:///D:/github/sam-app/amazonapi ; python_version >= "3.12"
-e file:///D:/github/sam-app/cognito ; python_version >= "3.12"
-e file:///D:/github/sam-app/db ; python_version >= "3.12"
-e file:///D:/github/sam-app/dzgrosecrets ; python_version >= "3.12"
-e file:///D:/github/sam-app/models ; python_version >= "3.12"
-e file:///D:/github/sam-app/razorpay ; python_version >= "3.12"
-e file:///D:/github/sam-app/storage ; python_version >= "3.12"
annotated-types==0.7.0 ; python_version >= "3.12"
anyio==4.9.0 ; python_version >= "3.12"
backoff==1.11.1 ; python_version >= "3.12"
boto3==1.39.13 ; python_version >= "3.12"
botocore==1.39.13 ; python_version >= "3.12"
certifi==2025.7.14 ; python_version >= "3.12"
cffi==1.17.1 ; python_version >= "3.12" and platform_python_implementation != "PyPy"
click==8.2.1 ; python_version >= "3.12"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.12"
cryptography==45.0.5 ; python_version >= "3.12"
dnspython==2.7.0 ; python_version >= "3.12"
email-validator==2.2.0 ; python_version >= "3.12"
fastapi-cli==0.0.8 ; python_version >= "3.12"
fastapi-cloud-cli==0.1.4 ; python_version >= "3.12"
fastapi==0.116.1 ; python_version >= "3.12"
h11==0.16.0 ; python_version >= "3.12"
httpcore==1.0.9 ; python_version >= "3.12"
httptools==0.6.4 ; python_version >= "3.12"
httpx==0.28.1 ; python_version >= "3.12"
idna==3.10 ; python_version >= "3.12"
iniconfig==2.1.0 ; python_version >= "3.12"
jinja2==3.1.6 ; python_version >= "3.12"
jmespath==1.0.1 ; python_version >= "3.12"
jwt==1.4.0 ; python_version >= "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.12"
markupsafe==3.0.2 ; python_version >= "3.12"
mdurl==0.1.2 ; python_version >= "3.12"
motor==3.7.1 ; python_version >= "3.12"
mypy-boto3-cognito-idp==1.39.0 ; python_version >= "3.12"
packaging==25.0 ; python_version >= "3.12"
pluggy==1.6.0 ; python_version >= "3.12"
pycparser==2.22 ; python_version >= "3.12" and platform_python_implementation != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.12"
pydantic==2.11.7 ; python_version >= "3.12"
pygments==2.19.2 ; python_version >= "3.12"
pyjwt==2.10.1 ; python_version >= "3.12"
pymongo==4.13.2 ; python_version >= "3.12"
pytest==8.4.1 ; python_version >= "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.12"
python-dotenv==1.1.1 ; python_version >= "3.12"
python-multipart==0.0.20 ; python_version >= "3.12"
pytz==2025.2 ; python_version >= "3.12"
pyyaml==6.0.2 ; python_version >= "3.12"
rich-toolkit==0.14.8 ; python_version >= "3.12"
rich==14.1.0 ; python_version >= "3.12"
rignore==0.6.4 ; python_version >= "3.12"
s3transfer==0.13.1 ; python_version >= "3.12"
sentry-sdk==2.33.2 ; python_version >= "3.12"
shellingham==1.5.4 ; python_version >= "3.12"
six==1.17.0 ; python_version >= "3.12"
sniffio==1.3.1 ; python_version >= "3.12"
starlette==0.47.2 ; python_version >= "3.12"
typer==0.16.0 ; python_version >= "3.12"
typing-extensions==4.14.1 ; python_version >= "3.12"
typing-inspection==0.4.1 ; python_version >= "3.12"
urllib3==2.5.0 ; python_version >= "3.12"
uvicorn==0.35.0 ; python_version >= "3.12"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.12"
watchfiles==1.1.0 ; python_version >= "3.12"
websockets==15.0.1 ; python_version >= "3.12"
